requirements.txt
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
slack_sdk>=3.27.0
